% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{compute_alpha_mu1}
\alias{compute_alpha_mu1}
\title{Calculates optimum alpha and mu for given other parameters}
\usage{
compute_alpha_mu1(pi, phi, r1, r2)
}
\arguments{
\item{pi}{Proportion of cases detected or “re-captured” by community surveillance}

\item{phi}{Proportion of contacts recalled}

\item{r1}{Ratio of cases not under surveillance (but contact traced) versus the cases under surveillance}

\item{r2}{Ratio of de novo cases verses detected cases under surveillance}
}
\value{
Optimum values of alpha and mu1
}
\description{
Calculates optimum alpha and mu for given other parameters
}
\examples{
compute_alpha_mu1(0.3, 0.7, 0.59, 1.5)
}
