% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{compute_phi_mu1}
\alias{compute_phi_mu1}
\title{Calculates optimum phi and mu for given other parameters}
\usage{
compute_phi_mu1(pi, alpha, r1, r2)
}
\arguments{
\item{pi}{Proportion of cases detected or “re-captured” by community surveillance}

\item{alpha}{Scaling of reproduction number for traced cases}

\item{r1}{Ratio of cases not under surveillance (but contact traced) versus the cases under surveillance}

\item{r2}{Ratio of de novo cases verses detected cases under surveillance}
}
\value{
Optimum values of pi and mu1
}
\description{
Calculates optimum phi and mu for given other parameters
}
\examples{
compute_phi_mu1(0.3, 0.7, 0.59, 1.5)
}
