% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{compute_pi_mu1}
\alias{compute_pi_mu1}
\title{Calculates optimum pi and mu for given other parameters}
\usage{
compute_pi_mu1(phi, alpha, r1, r2)
}
\arguments{
\item{phi}{Proportion of infected cases recalled as contacts}

\item{alpha}{Scaling of reproduction number for traced cases}

\item{r1}{Ratio of cases not under surveillance (but contact traced) versus the cases under surveillance}

\item{r2}{Ratio of de novo cases verses detected cases under surveillance}
}
\value{
Optimum values of pi and mu1
}
\description{
Calculates optimum pi and mu for given other parameters
}
\examples{
compute_pi_mu1(0.3, 0.7, 0.59, 1.5)
}
