% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGM.R
\name{define_NGM_four}
\alias{define_NGM_four}
\title{Defines the 3 state next generation matrix}
\usage{
define_NGM_four(pi, alpha, gamma, phi)
}
\arguments{
\item{pi}{Proportion of cases detected or “re-captured” by community surveillance}

\item{alpha}{Reduction in transmission intensity due to active surveillance}

\item{gamma}{Proportion of contacts actively under surveillance}

\item{phi}{Proportion of contacts recalled}
}
\value{
The four state next generation matrix
}
\description{
Defines the 3 state next generation matrix
}
\examples{
define_NGM_four(0.5, 0.8, 0.7, 0.5)
}
