% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution.R
\name{diff_funct_three_state}
\alias{diff_funct_three_state}
\title{Difference between two functional relationships of proportion of missing cases}
\usage{
diff_funct_three_state(pi, phi, alpha, r1, r2)
}
\arguments{
\item{pi}{Proportion of cases detected or “re-captured” by community surveillance}

\item{phi}{Proportion of contacts recalled}

\item{alpha}{Reduction in transmission intensity due to active surveillance}

\item{r1}{Ratio of cases not under surveillance (but contact traced) versus the cases under surveillance}

\item{r2}{Ratio of de novo cases verses detected cases under surveillance}
}
\value{
Difference between two functional relationships of proportion of missing cases
}
\description{
Difference between two functional relationships of proportion of missing cases
}
\examples{
diff_funct_three_state(0.3, 0.7, 0.47, 0.59, 1.5)
}
