% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGM.R
\name{prop_missing_cases}
\alias{prop_missing_cases}
\title{Calculates proportion of missing cases}
\usage{
prop_missing_cases(pi, alpha, gamma, phi)
}
\arguments{
\item{pi}{Proportion of cases detected or “re-captured” by community surveillance}

\item{alpha}{Reduction in transmission intensity due to active surveillance}

\item{gamma}{Proportion of contacts actively under surveillance}

\item{phi}{Proportion of contacts recalled}
}
\value{
Analytic proportion of missing cases
}
\description{
Calculates proportion of missing cases
}
\examples{
prop_missing_cases(0.5, 0.8, 0.7, 0.5)
}
